requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
python-decouple==3.8
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
mssql-django==1.3
pyodbc==4.0.39
